requirements.txt
flask
flask-cors
matplotlib
numpy
pandas
weasyprint
SQLAlchemy
psycopg2-binary
python-dotenv

